♻️ Refactor(my): my 도메인 패키지로 이주 #25 기타 반환타입 변경.
